#78811. Set of required changes. Phase III.

+++ src/Common/Productivity3DModels/src/Models/ProjectionParameter.cs
@@ -13,6 +13,6 @@
     /// <summary>
     /// The projection parameter's value.
     /// </summary>
-    public short Value { get; set; }
+    public double Value { get; set; }
   }
 }
